all tables are added in model package which are successfully executed and tables are verified currently using h2 db if db changed please check table connection view to ensure tables are connected withe each other perfectly

+++ src/main/java/com/omi/springboot_in_vs_code/controller/controller.java
@@ -32,7 +32,7 @@ public class Controller {
     }
 
     @GetMapping("/getbyid/{id}")
-    public Optional<Users> getUserById(@PathVariable int id) {
+    public Optional<Users> getUserById(@PathVariable Long id) {
         return userService.getUserById(id);
     }
 
@@ -48,7 +48,7 @@ public class Controller {
     }
 
     @DeleteMapping("/delete/{id}")
-    public String deleteUser(@PathVariable int id) {
+    public String deleteUser(@PathVariable Long id) {
         return userService.deleteUser(id);
     }
